+++ consensus/src/authority/actors/mod.rs
@@ -1,2 +1,10 @@
-pub mod orchestrator;
+pub mod resolver;
 pub mod voter;
+
+use crate::{authority::wire, Hash, Height, View};
+
+#[derive(Clone)]
+pub(crate) enum Proposal {
+    Reference(View, Height, Hash),
+    Populated(Hash, wire::Proposal),
+}
