--- consensus/src/authority/actors/orchestrator/mod.rs
@@ -1,2 +1,0 @@
-pub mod actor;
-pub mod ingress;
