+++ consensus/src/authority/actors/resolver/mod.rs
@@ -1,0 +1,5 @@
+mod actor;
+mod ingress;
+
+pub use actor::Actor;
+pub use ingress::{Mailbox, Message};
